Show lengths of large (more than 2G) files in directory mode.   lsl or ls=l or sort+s This use to fail on 32 bit systems.
Is it ok to set -D_FILE_OFFSET_BITS=64 all the time? Mac? Windows?
I haven't made the change to cmake yet.

+++ src/buffers.c
@@ -1543,7 +1543,7 @@ struct DSR {
 	int idx;
 	union {
 		time_t t;
-		size_t z;
+		off_t z;
 	} u;
 };
 static struct DSR *dsr_list;
